Close and-rewrite branch
This flavour really doesn't work so well. Any branch that returns an empty
result stream for one of its "and" branches will return no results. However,
there are lots of examples where there are good results to be had - eg

(and (job ?supervisor ?role)
     (unique (supervisor ?x ?supervisor)))

returns no results because the unique will always fail until the ?supervisor
variable has been bound on the other branch.

This *can* be fixed - if an and branch returns no results, then you can return a
single empty frame with some sort of filter on it but:
a) getting these filters to play nicely together (eg only provide filters where
they assign variables in a compatible manner) is quite hard

b) at that point we are basically going back to the same thing: using and
branches as filters, rather than as branches.